requirements.txt
matplotlib
python-pushover
pyyaml
seaborn
